+++ src/tooling/piral-cli/src/types/public.ts
@@ -219,7 +219,7 @@ export interface BundlerDefinition {
   buildPilet: BuildPiletBundlerDefinition;
 }
 
-export type PiletSchemaVersion = 'none' | 'v0' | 'v1' | 'v2';
+export type PiletSchemaVersion = 'none' | 'v0' | 'v1' | 'v2' | 'v3';
 
 export type SourceLanguage = 'js' | 'ts';