+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Updated `piral-debug-utils` to also work more seamlessly with `piral-base`
 - Updated `piral-cli` to have `pilet build` working without any Piral instances
 - Added variants of `piral-base` (minimal, Node.js, full)
+- Added support for new pilet schema `v3` (default remains at `v2`)
 
 ## 0.15.9 (April 11, 2023)